LPS-60972 This "jar" task configuration is specific to Liferay
